Remove wallet title
